+++ codec/codec_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package binary
+package codec
 
 import (
 	"bytes"
